refactor(ecs): move entity_manager.hpp to entity dir
