+++ core/CMakeLists.txt
@@ -70,7 +70,7 @@ set(CUBOS_CORE_SOURCE
     "src/cubos/core/al/oal_audio_device.cpp"
     "src/cubos/core/al/oal_audio_device.hpp"
 
-    "src/cubos/core/ecs/entity_manager.cpp"
+    "src/cubos/core/ecs/entity/manager.cpp"
     "src/cubos/core/ecs/component_manager.cpp"
     "src/cubos/core/ecs/commands.cpp"
     "src/cubos/core/ecs/blueprint.cpp"
