+++ core/include/cubos/core/ecs/entity/manager.hpp
