+++ core/include/cubos/core/ecs/storage.hpp
@@ -6,7 +6,7 @@
 
 #include <cubos/core/data/old/package.hpp>
 #include <cubos/core/data/old/serialization_map.hpp>
-#include <cubos/core/ecs/entity_manager.hpp>
+#include <cubos/core/ecs/entity/manager.hpp>
 
 namespace cubos::core::ecs
 {
