+++ engine/include/cubos/engine/collisions/broad_phase_collisions.hpp
@@ -8,7 +8,7 @@
 #include <unordered_set>
 #include <vector>
 
-#include <cubos/core/ecs/entity_manager.hpp>
+#include <cubos/core/ecs/entity/manager.hpp>
 
 using cubos::core::ecs::Entity;
